+++ packages/teradata/index.js
@@ -1,0 +1,5 @@
+import TeradataDriver from './TeradataDriver.js';
+export { default as TeradataDriver } from './TeradataDriver.js';
+export { default as TeradataConnection } from './TeradataConnection.js';
+export { default as TeradataStatement } from './TeradataStatement.js';
+export { default as TeradataPreparedStatement } from './TeradataPreparedStatement.js';